pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "ollama",
    "yfinance>=1.3.0",
    "apscheduler>=3.11.2",
    "aiohttp>=3.13.5",
    "tradingview-mcp-server",
    "langgraph>=1.1.8",
    "borsa-mcp",
    "mcp-server",
    "maverick-mcp",
    "kospi-kosdaq-stock-server",
    "aiofiles>=25.1.0",
    "jinja2>=3.1.6",
    "youtube-transcript-api>=1.2.4",
    "chromadb>=1.5.8",
    "protobuf<=3.20.3",
    "yt-dlp>=2026.3.17",
    "openai-whisper>=20250625",
]

[tool.uv.workspace]
members = [
    "temp_tools/tradingview-mcp",
    "temp_tools/borsa-mcp",
    "temp_tools/financial-datasets-mcp",
    "temp_tools/maverick-mcp",
    "temp_tools/kospi-kosdaq-stock-server",
]

[tool.uv.sources]
tradingview-mcp-server = { workspace = true }
borsa-mcp = { workspace = true }
mcp-server = { workspace = true }
maverick-mcp = { workspace = true }
kospi-kosdaq-stock-server = { workspace = true }


